feat: decompose engine.py and compiler.py into focused subpackages (Phase 4)
Split engine.py (1022 lines) into engine/ package with 7 focused modules
(validate, results, interrupts, run, resume, batch, tasks_dir) and
compiler.py (1189 lines) into compiler/ package with 7 modules (helpers,
routing, aggregation, execution, nodes, parallel, compile).

Re-export facades in __init__.py preserve backward compatibility for all
existing imports. Test patch targets updated to point at actual submodule
namespaces. Extracted shared handle_interrupts() to eliminate duplication
between run_flow and resume_flow.

Tasks T018–T035 completed. All 1379 tests pass.
