+++ src/fdsx/core/engine/__init__.py
@@ -1,0 +1,40 @@
+"""Engine package re-export facade.
+
+All public and tested-private symbols are re-exported here so that
+existing imports like ``from fdsx.core.engine import run_flow`` continue
+to work without changes.
+"""
+from .batch import run_batch
+from .resume import resume_flow
+from .results import (
+    _calc_elapsed,
+    _extract_results,
+    _find_failed_state,
+    _sanitize_state_for_log,
+)
+from .run import run_flow
+from .tasks_dir import (
+    _filter_actionable_entries,
+    _update_task_status,
+    _workflow_persist_id,
+    load_tasks_dir,
+    run_tasks_dir,
+)
+from .validate import FlowValidationError, validate_flow
+
+__all__ = [
+    "FlowValidationError",
+    "_calc_elapsed",
+    "_extract_results",
+    "_filter_actionable_entries",
+    "_find_failed_state",
+    "_sanitize_state_for_log",
+    "_update_task_status",
+    "_workflow_persist_id",
+    "load_tasks_dir",
+    "resume_flow",
+    "run_batch",
+    "run_flow",
+    "run_tasks_dir",
+    "validate_flow",
+]
