+++ tests/integration/test_split_spinner.py
@@ -131,10 +131,10 @@ class TestAutoSelectionSpinner:
             return workflows_dir / "test.yaml"
 
         with patch(
-            "fdsx.core.engine.resolve_workflow_for_task", side_effect=mock_resolve
+            "fdsx.core.selector.resolve_workflow_for_task", side_effect=mock_resolve
         ):
-            with patch("fdsx.core.engine.run_flow", return_value={"result": "ok"}):
-                with patch("fdsx.core.engine.display_tasks_dir_summary"):
+            with patch("fdsx.core.engine.tasks_dir.run_flow", return_value={"result": "ok"}):
+                with patch("fdsx.core.engine.tasks_dir.display_tasks_dir_summary"):
                     engine.run_tasks_dir(
                         None,
                         tasks_dir,
@@ -175,13 +175,13 @@ class TestAutoSelectionSpinner:
 
         _MockSpinner.reset()
 
-        with patch("fdsx.core.engine.Spinner", side_effect=_MockSpinner):
+        with patch("fdsx.core.engine.tasks_dir.Spinner", side_effect=_MockSpinner):
             with patch(
-                "fdsx.core.engine.resolve_workflow_for_task",
+                "fdsx.core.selector.resolve_workflow_for_task",
                 return_value=workflows_dir / "test.yaml",
             ):
-                with patch("fdsx.core.engine.run_flow", return_value={"result": "ok"}):
-                    with patch("fdsx.core.engine.display_tasks_dir_summary"):
+                with patch("fdsx.core.engine.tasks_dir.run_flow", return_value={"result": "ok"}):
+                    with patch("fdsx.core.engine.tasks_dir.display_tasks_dir_summary"):
                         engine.run_tasks_dir(
                             None,
                             tasks_dir,
@@ -229,9 +229,9 @@ class TestAutoSelectionSpinner:
 
         _MockSpinner.reset()
 
-        with patch("fdsx.core.engine.Spinner", side_effect=_MockSpinner):
-            with patch("fdsx.core.engine.run_flow", return_value={"result": "ok"}):
-                with patch("fdsx.core.engine.display_tasks_dir_summary"):
+        with patch("fdsx.core.engine.tasks_dir.Spinner", side_effect=_MockSpinner):
+            with patch("fdsx.core.engine.tasks_dir.run_flow", return_value={"result": "ok"}):
+                with patch("fdsx.core.engine.tasks_dir.display_tasks_dir_summary"):
                     engine.run_tasks_dir(
                         None,
                         tasks_dir,
@@ -269,9 +269,9 @@ class TestAutoSelectionSpinner:
 
         _MockSpinner.reset()
 
-        with patch("fdsx.core.engine.Spinner", side_effect=_MockSpinner):
-            with patch("fdsx.core.engine.run_flow", return_value={"result": "ok"}):
-                with patch("fdsx.core.engine.display_tasks_dir_summary"):
+        with patch("fdsx.core.engine.tasks_dir.Spinner", side_effect=_MockSpinner):
+            with patch("fdsx.core.engine.tasks_dir.run_flow", return_value={"result": "ok"}):
+                with patch("fdsx.core.engine.tasks_dir.display_tasks_dir_summary"):
                     engine.run_tasks_dir(
                         workflow_path,
                         tasks_dir,
