+++ tests/integration/test_tasks_dir.py
@@ -176,8 +176,8 @@ class TestRunTasksDir:
             tf2 = TaskFile(entries=[TaskEntry(description="task B")])
             save_task_file(tasks_dir / "002-b.yaml", tf2)
 
-            with patch("fdsx.core.engine.run_flow", return_value={"result": "ok"}):
-                with patch("fdsx.core.engine.display_tasks_dir_summary"):
+            with patch("fdsx.core.engine.tasks_dir.run_flow", return_value={"result": "ok"}):
+                with patch("fdsx.core.engine.tasks_dir.display_tasks_dir_summary"):
                     results = engine.run_tasks_dir(
                         flow_path, tasks_dir, auto_workflow=True
                     )
@@ -217,8 +217,8 @@ class TestRunTasksDir:
                 run_count[0] += 1
                 return {"result": "ok"}
 
-            with patch("fdsx.core.engine.run_flow", side_effect=mock_run_flow):
-                with patch("fdsx.core.engine.display_tasks_dir_summary"):
+            with patch("fdsx.core.engine.tasks_dir.run_flow", side_effect=mock_run_flow):
+                with patch("fdsx.core.engine.tasks_dir.display_tasks_dir_summary"):
                     results = engine.run_tasks_dir(
                         flow_path, tasks_dir, auto_workflow=True
                     )
@@ -247,8 +247,8 @@ class TestRunTasksDir:
             )
             save_task_file(tasks_dir / "001-test.yaml", tf)
 
-            with patch("fdsx.core.engine.run_flow", return_value={"result": "ok"}):
-                with patch("fdsx.core.engine.display_tasks_dir_summary"):
+            with patch("fdsx.core.engine.tasks_dir.run_flow", return_value={"result": "ok"}):
+                with patch("fdsx.core.engine.tasks_dir.display_tasks_dir_summary"):
                     results = engine.run_tasks_dir(
                         flow_path, tasks_dir, auto_workflow=True
                     )
@@ -278,8 +278,8 @@ class TestRunTasksDir:
             )
             save_task_file(tasks_dir / "001-test.yaml", tf)
 
-            with patch("fdsx.core.engine.run_flow", return_value={"result": "ok"}):
-                with patch("fdsx.core.engine.display_tasks_dir_summary"):
+            with patch("fdsx.core.engine.tasks_dir.run_flow", return_value={"result": "ok"}):
+                with patch("fdsx.core.engine.tasks_dir.display_tasks_dir_summary"):
                     results = engine.run_tasks_dir(
                         flow_path, tasks_dir, auto_workflow=True
                     )
@@ -308,8 +308,8 @@ class TestRunTasksDir:
                 call_count[0] += 1
                 return {"result": "ok"}
 
-            with patch("fdsx.core.engine.run_flow", side_effect=mock_run_flow):
-                with patch("fdsx.core.engine.display_tasks_dir_summary"):
+            with patch("fdsx.core.engine.tasks_dir.run_flow", side_effect=mock_run_flow):
+                with patch("fdsx.core.engine.tasks_dir.display_tasks_dir_summary"):
                     results = engine.run_tasks_dir(
                         flow_path, tasks_dir, auto_workflow=True
                     )
@@ -343,9 +343,9 @@ class TestRunTasksDir:
                     raise RuntimeError("Task 1 failed")
                 return {"result": "ok"}
 
-            with patch("fdsx.core.engine.run_flow", side_effect=mock_run_flow):
-                with patch("fdsx.core.engine.display_tasks_dir_summary"):
-                    with patch("fdsx.core.engine.input", side_effect=["y"]):
+            with patch("fdsx.core.engine.tasks_dir.run_flow", side_effect=mock_run_flow):
+                with patch("fdsx.core.engine.tasks_dir.display_tasks_dir_summary"):
+                    with patch("fdsx.core.engine.tasks_dir.input", side_effect=["y"]):
                         results = engine.run_tasks_dir(
                             flow_path, tasks_dir, auto_workflow=True
                         )
@@ -379,9 +379,9 @@ class TestRunTasksDir:
                     raise RuntimeError("Task 1 failed")
                 return {"result": "ok"}
 
-            with patch("fdsx.core.engine.run_flow", side_effect=mock_run_flow):
-                with patch("fdsx.core.engine.display_tasks_dir_summary"):
-                    with patch("fdsx.core.engine.input", side_effect=["n"]):
+            with patch("fdsx.core.engine.tasks_dir.run_flow", side_effect=mock_run_flow):
+                with patch("fdsx.core.engine.tasks_dir.display_tasks_dir_summary"):
+                    with patch("fdsx.core.engine.tasks_dir.input", side_effect=["n"]):
                         results = engine.run_tasks_dir(
                             flow_path, tasks_dir, auto_workflow=True
                         )
@@ -402,9 +402,9 @@ class TestRunTasksDir:
             save_task_file(tasks_dir / "001-done.yaml", tf)
 
             with patch(
-                "fdsx.core.engine.run_flow", return_value={"result": "ok"}
+                "fdsx.core.engine.tasks_dir.run_flow", return_value={"result": "ok"}
             ) as mock_run:
-                with patch("fdsx.core.engine.display_tasks_dir_summary"):
+                with patch("fdsx.core.engine.tasks_dir.display_tasks_dir_summary"):
                     results = engine.run_tasks_dir(
                         flow_path, tasks_dir, auto_workflow=True
                     )
@@ -429,9 +429,9 @@ class TestRunTasksDir:
             save_task_file(tasks_dir / "001-done.yaml", tf)
 
             with patch(
-                "fdsx.core.engine.run_flow", return_value={"result": "ok"}
+                "fdsx.core.engine.tasks_dir.run_flow", return_value={"result": "ok"}
             ) as mock_run:
-                with patch("fdsx.core.engine.display_tasks_dir_summary"):
+                with patch("fdsx.core.engine.tasks_dir.display_tasks_dir_summary"):
                     results = engine.run_tasks_dir(
                         flow_path, tasks_dir, auto_workflow=True
                     )
@@ -463,9 +463,9 @@ class TestRunTasksDir:
             def mock_run_flow(flow_path, inputs, thread_id, base_dir, **kwargs):
                 raise RuntimeError("Failed again")
 
-            with patch("fdsx.core.engine.run_flow", side_effect=mock_run_flow):
-                with patch("fdsx.core.engine.display_tasks_dir_summary"):
-                    with patch("fdsx.core.engine.input", side_effect=["n"]):
+            with patch("fdsx.core.engine.tasks_dir.run_flow", side_effect=mock_run_flow):
+                with patch("fdsx.core.engine.tasks_dir.display_tasks_dir_summary"):
+                    with patch("fdsx.core.engine.tasks_dir.input", side_effect=["n"]):
                         results = engine.run_tasks_dir(
                             flow_path, tasks_dir, auto_workflow=True
                         )
@@ -489,9 +489,9 @@ class TestRunTasksDir:
             def mock_run_flow(flow_path, inputs, thread_id, base_dir, **kwargs):
                 raise RuntimeError("Task failed")
 
-            with patch("fdsx.core.engine.run_flow", side_effect=mock_run_flow):
-                with patch("fdsx.core.engine.display_tasks_dir_summary"):
-                    with patch("fdsx.core.engine.input", side_effect=["n"]):
+            with patch("fdsx.core.engine.tasks_dir.run_flow", side_effect=mock_run_flow):
+                with patch("fdsx.core.engine.tasks_dir.display_tasks_dir_summary"):
+                    with patch("fdsx.core.engine.tasks_dir.input", side_effect=["n"]):
                         engine.run_tasks_dir(flow_path, tasks_dir, auto_workflow=True)
 
             loaded = load_task_file(tasks_dir / "001-test.yaml")
@@ -556,8 +556,8 @@ class TestTasksDirCli:
         tf = TaskFile(entries=[TaskEntry(description="cli task")])
         save_task_file(tasks_dir / "001-test.yaml", tf)
 
-        with patch("fdsx.core.engine.run_flow", return_value={"result": "ok"}):
-            with patch("fdsx.core.engine.display_tasks_dir_summary"):
+        with patch("fdsx.core.engine.tasks_dir.run_flow", return_value={"result": "ok"}):
+            with patch("fdsx.core.engine.tasks_dir.display_tasks_dir_summary"):
                 runner = CliRunner()
                 result = runner.invoke(
                     app,
@@ -656,8 +656,8 @@ class TestMoveToCompletedOnRunTasksDir:
             tf = TaskFile(entries=[TaskEntry(description="task A")])
             save_task_file(tasks_dir / "001-a.yaml", tf)
 
-            with patch("fdsx.core.engine.run_flow", return_value={"result": "ok"}):
-                with patch("fdsx.core.engine.display_tasks_dir_summary"):
+            with patch("fdsx.core.engine.tasks_dir.run_flow", return_value={"result": "ok"}):
+                with patch("fdsx.core.engine.tasks_dir.display_tasks_dir_summary"):
                     engine.run_tasks_dir(flow_path, tasks_dir, auto_workflow=True)
 
             assert not (tasks_dir / "001-a.yaml").exists()
@@ -671,9 +671,9 @@ class TestMoveToCompletedOnRunTasksDir:
             tf = TaskFile(entries=[TaskEntry(description="task A")])
             save_task_file(tasks_dir / "001-a.yaml", tf)
 
-            with patch("fdsx.core.engine.run_flow", side_effect=RuntimeError("fail")):
-                with patch("fdsx.core.engine.display_tasks_dir_summary"):
-                    with patch("fdsx.core.engine.input", side_effect=["n"]):
+            with patch("fdsx.core.engine.tasks_dir.run_flow", side_effect=RuntimeError("fail")):
+                with patch("fdsx.core.engine.tasks_dir.display_tasks_dir_summary"):
+                    with patch("fdsx.core.engine.tasks_dir.input", side_effect=["n"]):
                         engine.run_tasks_dir(flow_path, tasks_dir, auto_workflow=True)
 
             # File with failed entry must remain in tasks_dir
@@ -702,9 +702,9 @@ class TestMoveToCompletedOnRunTasksDir:
                     raise RuntimeError("first fails")
                 return {"result": "ok"}
 
-            with patch("fdsx.core.engine.run_flow", side_effect=mock_run_flow):
-                with patch("fdsx.core.engine.display_tasks_dir_summary"):
-                    with patch("fdsx.core.engine.input", side_effect=["y"]):
+            with patch("fdsx.core.engine.tasks_dir.run_flow", side_effect=mock_run_flow):
+                with patch("fdsx.core.engine.tasks_dir.display_tasks_dir_summary"):
+                    with patch("fdsx.core.engine.tasks_dir.input", side_effect=["y"]):
                         engine.run_tasks_dir(flow_path, tasks_dir, auto_workflow=True)
 
             assert (tasks_dir / "001-mixed.yaml").exists()
@@ -721,8 +721,8 @@ class TestMoveToCompletedOnRunTasksDir:
             )
             save_task_file(tasks_dir / "001-done.yaml", tf)
 
-            with patch("fdsx.core.engine.run_flow") as mock_run:
-                with patch("fdsx.core.engine.display_tasks_dir_summary"):
+            with patch("fdsx.core.engine.tasks_dir.run_flow") as mock_run:
+                with patch("fdsx.core.engine.tasks_dir.display_tasks_dir_summary"):
                     engine.run_tasks_dir(flow_path, tasks_dir, auto_workflow=True)
 
             mock_run.assert_not_called()
@@ -737,10 +737,10 @@ class TestMoveToCompletedOnRunTasksDir:
             tf = TaskFile(entries=[TaskEntry(description="task A")])
             save_task_file(tasks_dir / "001-a.yaml", tf)
 
-            with patch("fdsx.core.engine.run_flow", return_value={"result": "ok"}):
-                with patch("fdsx.core.engine.display_tasks_dir_summary"):
+            with patch("fdsx.core.engine.tasks_dir.run_flow", return_value={"result": "ok"}):
+                with patch("fdsx.core.engine.tasks_dir.display_tasks_dir_summary"):
                     with patch(
-                        "fdsx.core.engine.move_task_to_completed",
+                        "fdsx.core.engine.tasks_dir.move_task_to_completed",
                         side_effect=OSError("disk full"),
                     ):
                         results = engine.run_tasks_dir(
@@ -767,8 +767,8 @@ class TestMoveToCompletedOnRunTasksDir:
             tf = TaskFile(entries=[TaskEntry(description="task A")])
             save_task_file(tasks_dir / "001-a.yaml", tf)
 
-            with patch("fdsx.core.engine.run_flow", return_value={"result": "ok"}):
-                with patch("fdsx.core.engine.display_tasks_dir_summary"):
+            with patch("fdsx.core.engine.tasks_dir.run_flow", return_value={"result": "ok"}):
+                with patch("fdsx.core.engine.tasks_dir.display_tasks_dir_summary"):
                     results = engine.run_tasks_dir(
                         flow_path, tasks_dir, auto_workflow=True
                     )
@@ -825,8 +825,8 @@ class TestBatchEditFlow:
         mock_assignments = {(0, 0): wf_path}
 
         with patch("fdsx.core.selector.get_provider", return_value=MagicMock()):
-            with patch("fdsx.core.engine.run_flow", return_value={"result": "ok"}):
-                with patch("fdsx.core.engine.display_tasks_dir_summary"):
+            with patch("fdsx.core.engine.tasks_dir.run_flow", return_value={"result": "ok"}):
+                with patch("fdsx.core.engine.tasks_dir.display_tasks_dir_summary"):
                     with patch(
                         "fdsx.display.terminal.confirm_workflow_assignments_interactive",
                         return_value=mock_assignments,
@@ -855,7 +855,7 @@ class TestBatchEditFlow:
         save_task_file(tasks_dir / "001-test.yaml", tf)
 
         with patch("fdsx.core.selector.get_provider", return_value=MagicMock()):
-            with patch("fdsx.core.engine.display_tasks_dir_summary"):
+            with patch("fdsx.core.engine.tasks_dir.display_tasks_dir_summary"):
                 with patch(
                     "fdsx.display.terminal.confirm_workflow_assignments_interactive",
                     return_value=None,
@@ -884,8 +884,8 @@ class TestBatchEditFlow:
         save_task_file(tasks_dir / "001-test.yaml", tf)
 
         with patch("fdsx.core.selector.get_provider", return_value=MagicMock()):
-            with patch("fdsx.core.engine.run_flow", return_value={"result": "ok"}):
-                with patch("fdsx.core.engine.display_tasks_dir_summary"):
+            with patch("fdsx.core.engine.tasks_dir.run_flow", return_value={"result": "ok"}):
+                with patch("fdsx.core.engine.tasks_dir.display_tasks_dir_summary"):
                     with patch(
                         "fdsx.display.terminal.confirm_workflow_assignments_interactive"
                     ) as mock_cui:
@@ -909,8 +909,8 @@ class TestRunTasksDirQuietFlagPropagation:
             tf = TaskFile(entries=[TaskEntry(description="task A")])
             save_task_file(tasks_dir / "001-a.yaml", tf)
 
-            with patch("fdsx.core.engine.run_flow") as mock_run_flow:
-                with patch("fdsx.core.engine.display_tasks_dir_summary"):
+            with patch("fdsx.core.engine.tasks_dir.run_flow") as mock_run_flow:
+                with patch("fdsx.core.engine.tasks_dir.display_tasks_dir_summary"):
                     engine.run_tasks_dir(
                         flow_path, tasks_dir, auto_workflow=True, quiet=True
                     )
@@ -927,8 +927,8 @@ class TestRunTasksDirQuietFlagPropagation:
             tf = TaskFile(entries=[TaskEntry(description="task A")])
             save_task_file(tasks_dir / "001-a.yaml", tf)
 
-            with patch("fdsx.core.engine.run_flow") as mock_run_flow:
-                with patch("fdsx.core.engine.display_tasks_dir_summary"):
+            with patch("fdsx.core.engine.tasks_dir.run_flow") as mock_run_flow:
+                with patch("fdsx.core.engine.tasks_dir.display_tasks_dir_summary"):
                     engine.run_tasks_dir(flow_path, tasks_dir, auto_workflow=True)
 
             assert mock_run_flow.called
@@ -954,8 +954,8 @@ class TestRunTasksDirSourceInjection:
                 captured_inputs.append(dict(inputs))
                 return {"result": "ok"}
 
-            with patch("fdsx.core.engine.run_flow", side_effect=mock_run_flow):
-                with patch("fdsx.core.engine.display_tasks_dir_summary"):
+            with patch("fdsx.core.engine.tasks_dir.run_flow", side_effect=mock_run_flow):
+                with patch("fdsx.core.engine.tasks_dir.display_tasks_dir_summary"):
                     engine.run_tasks_dir(flow_path, tasks_dir, auto_workflow=True)
 
             assert len(captured_inputs) == 1
@@ -976,8 +976,8 @@ class TestRunTasksDirSourceInjection:
                 captured_inputs.append(dict(inputs))
                 return {"result": "ok"}
 
-            with patch("fdsx.core.engine.run_flow", side_effect=mock_run_flow):
-                with patch("fdsx.core.engine.display_tasks_dir_summary"):
+            with patch("fdsx.core.engine.tasks_dir.run_flow", side_effect=mock_run_flow):
+                with patch("fdsx.core.engine.tasks_dir.display_tasks_dir_summary"):
                     engine.run_tasks_dir(flow_path, tasks_dir, auto_workflow=True)
 
             assert len(captured_inputs) == 1
